moved grant/revoke/check role from user- to role-service
