+++ core/src/main/java/org/openstack4j/api/identity/RoleService.java
@@ -12,15 +12,65 @@ import org.openstack4j.model.identity.Role;
  */
 public interface RoleService extends RestService {
 
-	/**
-	 * Adds a project based role to a user
-	 *
-	 * @param projectId the project id
-	 * @param userId the user id
-	 * @param roleId the role id
-	 * @return the action response
-	 */
-	ActionResponse addRoleToUserInProject(String projectId, String userId, String roleId);
+    /**
+     * grants a role to a specified user in project context
+     *
+     * @param projectId the project id
+     * @param userId the user id
+     * @param roleId the role id
+     * @return the action response
+     */
+    ActionResponse grantProjectUserRole(String projectId, String userId, String roleId);
+
+    /**
+     * revokes a role to a specified user in project context
+     *
+     * @param projectId the project id
+     * @param userId the user id
+     * @param roleId the role id
+     * @return the action response
+     */
+    ActionResponse revokeProjectUserRole(String projectId, String userId, String roleId);
+
+    /**
+     * checks if a user has a specific role in a given project-context
+     *
+     * @param projectId the project id
+     * @param userId the user id
+     * @param roleId the role id
+     * @return the ActionResponse
+     */
+    ActionResponse checkProjectUserRole(String projectId, String userId, String roleId);
+
+    /**
+     * grants a role to a specified user in domain context
+     *
+     * @param domainId the domain id
+     * @param userId the user id
+     * @param roleId the role id
+     * @return the action response
+     */
+    ActionResponse grantDomainUserRole(String domainId, String userId, String roleId);
+
+    /**
+     * revokes a role to a specified user in domain context
+     *
+     * @param domainId the domain id
+     * @param userId the user id
+     * @param roleId the role id
+     * @return the action response
+     */
+    ActionResponse revokeDomainUserRole(String domainId, String userId, String roleId);
+
+    /**
+     * checks if a user has a specific role in a given domain-context
+     *
+     * @param domainId the domain id
+     * @param userId the user id
+     * @param roleId the role id
+     * @return the ActionResponse
+     */
+    ActionResponse checkDomainUserRole(String domainId, String userId, String roleId);
 
 	/**
 	 * Lists the global roles
