+++ core/src/main/java/org/openstack4j/openstack/identity/internal/ProjectServiceImpl.java
@@ -8,12 +8,15 @@ import org.openstack4j.model.identity.Project;
 import org.openstack4j.openstack.identity.domain.KeystoneProject;
 import org.openstack4j.openstack.internal.BaseOpenStackService;
 
-public class ProjectServiceImpl extends BaseOpenStackService implements ProjectService  {
+public class ProjectServiceImpl extends BaseOpenStackService implements ProjectService {
 
-	@Override
-	public Project create(Project project) {
-		checkNotNull(project);
-		return post(KeystoneProject.class, PATH_PROJECTS).entity(project).execute();
-	}
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public Project create(Project project) {
+        checkNotNull(project);
+        return post(KeystoneProject.class, PATH_PROJECTS).entity(project).execute();
+    }
 
 }
