+++ core/src/main/java/org/openstack4j/openstack/identity/internal/RoleServiceImpl.java
@@ -5,7 +5,6 @@ import static com.google.common.base.Preconditions.checkNotNull;
 import java.util.List;
 
 import org.openstack4j.api.identity.RoleService;
-import org.openstack4j.core.transport.HttpMethod;
 import org.openstack4j.model.compute.ActionResponse;
 import org.openstack4j.model.identity.Role;
 import org.openstack4j.openstack.identity.domain.KeystoneRole.Roles;
@@ -17,34 +16,88 @@ import org.openstack4j.openstack.internal.BaseOpenStackService;
  */
 public class RoleServiceImpl extends BaseOpenStackService implements RoleService {
 
-	/**
-	 * {@inheritDoc}
-	 */
-	@Override
-	public ActionResponse addRoleToUserInProject(String projectId, String userId, String roleId) {
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public ActionResponse grantProjectUserRole(String projectId, String userId, String roleId) {
+        checkNotNull(userId);
+        checkNotNull(projectId);
+        checkNotNull(roleId);
+        return put(ActionResponse.class, uri("projects/%s/users/%s/roles/%s", projectId, userId, roleId)).execute();
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public ActionResponse revokeProjectUserRole(String projectId, String userId, String roleId) {
+        checkNotNull(userId);
+        checkNotNull(projectId);
+        checkNotNull(roleId);
+        return deleteWithResponse(uri("projects/%s/users/%s/roles/%s", projectId, userId, roleId)).execute();
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public ActionResponse checkProjectUserRole(String projectId, String userId, String roleId) {
         checkNotNull(projectId);
         checkNotNull(userId);
         checkNotNull(roleId);
-        String uri = uri("/projects/%s/users/%s/roles/%s", projectId, userId, roleId);
-        return request(HttpMethod.PUT, ActionResponse.class, uri).execute();
-	}
+        return head(ActionResponse.class, uri("/projects/%s/users/%s/roles/%s", projectId, userId, roleId)).execute();
+    }
 
-	/**
-	 * {@inheritDoc}
-	 */
-	@Override
-	public List<? extends Role> list() {
-		return get(Roles.class, uri("/roles")).execute().getList();
-	}
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public ActionResponse grantDomainUserRole(String domainId, String userId, String roleId) {
+        checkNotNull(userId);
+        checkNotNull(domainId);
+        checkNotNull(roleId);
+        return put(ActionResponse.class, uri("domains/%s/users/%s/roles/%s", domainId, userId, roleId)).execute();
+    }
 
-	/**
-	 * {@inheritDoc}
-	 */
-	@Override
-	public List<? extends Role> getByName(String name) {
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public ActionResponse revokeDomainUserRole(String domainId, String userId, String roleId) {
+        checkNotNull(userId);
+        checkNotNull(domainId);
+        checkNotNull(roleId);
+        return deleteWithResponse(uri("domains/%s/users/%s/roles/%s", domainId, userId, roleId)).execute();
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public ActionResponse checkDomainUserRole(String domainId, String userId, String roleId) {
+        checkNotNull(domainId);
+        checkNotNull(userId);
+        checkNotNull(roleId);
+        return head(ActionResponse.class, uri("/domains/%s/users/%s/roles/%s", domainId, userId, roleId)).execute();
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public List<? extends Role> list() {
+        return get(Roles.class, uri("/roles")).execute().getList();
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public List<? extends Role> getByName(String name) {
         checkNotNull(name);
         return get(Roles.class, "/roles").param("name", name).execute().getList();
 
-	}
+    }
 
 }
